Trying to get Red-Black deletion to work.
I'm trying to get the delete method for RBT to work correctly.
I've been trying different tactics, but it looks as if just removing
the prune method from the RBT deletion is resulting in the correct
tree form. I'm about to convert everything to read in integers so that
I can easily test the tree with an RBT visualization in order to
see if my tree really is correct.

Side Notes:
    Code still needs to be cleaned up.
    I've corrected the niece and nephew methods to work for the
    deletion pseudocode.
    I've also added some if checks for empty trees and not being
    able to find nodes, just in case that happens.